# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  controller.ReviewCreateRequest:
    properties:
      content:
        type: string
      images:
        type: string
      order_id:
        type: integer
      rating:
        type: integer
      room_id:
        type: integer
      status:
        default: 1
        type: integer
      user_id:
        type: integer
    required:
    - order_id
    - rating
    - room_id
    type: object
...
  model.Review:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      images:
        type: string
      order_id:
        type: integer
      rating:
        type: integer
      room_id:
        type: integer
      status:
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/model.User'
      user_id:
        type: integer
    type: object
...
  model.User:
    properties:
      avatar:
        type: string
      created_at:
        type: string
      gender:
        type: integer
      id:
        type: integer
      nickname:
        type: string
      open_id:
        type: string
      phone:
        type: string
      realname:
        type: string
      status:
        type: integer
      updated_at:
        type: string
    type: object
  response.Response:
    properties:
      code:
        description: 响应状态码
        type: integer
      data:
        description: 响应数据，omitempty表示为空时不序列化
      message:
        description: 响应消息
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 棋室管理系统后端 API 文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: 棋室管理系统 API
  version: "1.0"
paths:
...
  /reviews:
    get:
      consumes:
      - application/json
      description: 获取评价列表，支持按房间ID筛选
      parameters:
      - description: 房间ID
        in: query
        name: room_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Review'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取评价列表
      tags:
      - 评价
    post:
      consumes:
      - application/json
      description: 用户创建评价，管理员可指定用户ID和状态
      parameters:
      - description: 评价信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller.ReviewCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Review'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 创建评价
      tags:
      - 评价
  /reviews/{id}:
    delete:
      consumes:
      - application/json
      description: 管理员删除评价
      parameters:
      - description: 评价ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除评价
      tags:
      - 评价
    get:
      consumes:
      - application/json
      description: 根据评价ID获取评价详情
      parameters:
      - description: 评价ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Review'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取评价详情
      tags:
      - 评价
    put:
      consumes:
      - application/json
      description: 管理员更新评价
      parameters:
      - description: 评价ID
        in: path
        name: id
        required: true
        type: string
      - description: 评价信息
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Review'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 更新评价
      tags:
      - 评价
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"